Partially fix filter control issues with Shiny; add Shiny docs
The filter controls don't work great with Shiny when they are rendered
in uiOutput. (It's hard not to render them in uiOutput because you don't
generally have the SharedData when constructing the initial UI.) Mostly
the problem is that renderUI obliterates the current state of the
filter. There was another problem here that filter controls were not
getting bound/initialized after the first renderUI execution, but that
should be fixed in this commit.

Shiny docs are about 80% complete.