Fix #40: SharedData across shiny modules
SharedData group should transcend modules (it's a global
namespace). This isn't as crazy as it sounds, because
the default group value is a random ID.

`SharedData$new(..., group = session$ns("groupname"))` can
be used to opt into namespaces if that's desired.

Note that this change doesn't break a scenario that works
today; before this commit, the reactive accessors on
SharedData that originated in modules, simply did not work.